requirements.txt
playwright==1.57.0
pytest==9.0.2
pytest-playwright==0.7.2
pytest-html==4.1.1
allure-pytest==2.15.3
allure-python-commons

